test: continue config flow user test split
